Add oneOf, anyOf validations
